\name{modelSoftwareMVBSN-package}
\alias{modelSoftwareMVBSN-package}
\alias{modelSoftwareMVBSN}
\docType{package}
\title{
  An R Package for Bayesian social network modelling.
}
\description{
  An R Package for Bayesian social network modelling.
}
\details{
  An R Package for Bayesian social network modelling.
}
\author{
George Gerogiannis

Maintainer: George Gerogiannis <g.gerogiannis.1@research.gla.ac.uk>
}
\keyword{ package }
