\name{plot.modelSoftwareMVBSN}
\alias{plot.modelSoftwareMVBSN}

\title{
  A function that plots visual MCMC diagnostics of the fitted model.
}

\description{
  This function plots visual MCMC diagnostics of the fitted model.
}

\usage{
  \method{plot}{modelSoftwareMVBSN}(x, \dots)
}

\arguments{

  \item{x}{
    Type: data.frame.
  }
  
  \item{\dots}{
    Ignored.s
  }
  
}

\author{
George Gerogiannis
}
