\name{plotRandomEffects}
\alias{plotRandomEffects}

\title{
  A function that plots the true random effect values against the posterior means of the random effects for the fitted model.
}

\description{
  This function plots the true random effect values against the posterior means of the random effects for the fitted model.
}

\usage{
  plotRandomEffects(randomEffectsSamples, trueRandomEffects)
}

\arguments{

  \item{randomEffectsSamples}{
    Type: data.frame.
  }
  
  \item{trueRandomEffects}{
    Type: vector.
  }
  
}


\author{
George Gerogiannis
}
