\name{summary.modelSoftwareMVBSN}
\alias{summary.modelSoftwareMVBSN}

\title{
  A function that gets a summary of the fitted model.
}

\description{
  This function gets a summary of the fitted model.
}

\usage{
  \method{summary}{modelSoftwareMVBSN}(object, \dots)
}

\arguments{

  \item{object}{
    Type: List.
  }
  
  \item{\dots}{
    Ignored.s
  }
  
}

\author{
George Gerogiannis
}
